hub-backend/requirements.txt
dj-rest-auth
django-allauth
django-cors-headers 